core/requirements.txt
# SupaBrain Core - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25

# Vector Search & Embeddings
sentence-transformers==2.3.1
pgvector==0.2.4
numpy==1.26.3

# ML & NLP (CPU-only - sentence-transformers pulls compatible torch)
transformers==4.37.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
